+++ projects/monitor-subscription/src/lib/modules/monitor-subscription.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ResultsComponent } from '../../public-api';
+
+@NgModule({
+  imports: [ResultsComponent, CommonModule],
+  exports: [ResultsComponent]
+})
+export class MonitorSubscriptionModule { }