+++ projects/monitor-subscription/src/public-api.ts
@@ -8,3 +8,4 @@ export * from './lib/services/logger.service';
 export * from './lib/services/destroy-subscription.service';
 export * from './lib/tools/active-subscription';
 export * from './lib/components/results/results.component'
+export * from './lib/modules/monitor-subscription.module'